backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
psycopg2-binary==2.9.9
redis==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# ML & Data Science
scikit-learn==1.3.2
numpy==1.24.4
pandas==2.1.4
joblib==1.3.2
scipy==1.11.4

# Advanced ML algorithms
GPy==1.12.0
chaospy==4.3.7
torch==2.1.1
tensorflow==2.15.0
xgboost==2.0.2
lightgbm==4.1.0

# Hyperparameter optimization
optuna==3.5.0
scikit-optimize==0.9.0

# Model validation and metrics
mlflow==2.8.1

# Task Queue
celery==5.3.4

# File handling
python-minio==7.2.0
aiofiles==23.2.1

# Monitoring & Logging
prometheus-client==0.19.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1